优化 app\services\analysis\__init__.py

+++ app/services/analysis/models.py
@@ -6,6 +6,7 @@
 from dataclasses import dataclass
 from typing import List
 from datetime import date
+from decimal import Decimal
 
 @dataclass
 class Period:
@@ -17,45 +18,45 @@ class Period:
 @dataclass
 class CoreMetrics:
     """核心现金流指标"""
-    current_total_assets: float
-    total_income: float
-    total_expense: float
-    net_income: float
-    income_change_percentage: float
-    expense_change_percentage: float
-    net_change_percentage: float
-    emergency_reserve_months: float
+    current_total_assets: Decimal
+    total_income: Decimal
+    total_expense: Decimal
+    net_income: Decimal
+    income_change_percentage: Decimal
+    expense_change_percentage: Decimal
+    net_change_percentage: Decimal
+    emergency_reserve_months: Decimal
 
 @dataclass
 class CompositionItem:
     """构成项目（收入或支出分类）"""
     name: str
-    amount: float
-    percentage: float
+    amount: Decimal
+    percentage: Decimal
     count: int
 
 @dataclass
 class TrendPoint:
     """趋势数据点"""
     date: str
-    value: float
+    value: Decimal
 
 @dataclass
 class TopExpenseCategory:
     """支出分类排行项目"""
     category: str
-    total_amount: float
-    percentage: float
+    total_amount: Decimal
+    percentage: Decimal
     count: int
 
 @dataclass
 class RecurringExpense:
     """周期性支出项目"""
     category: str
-    total_amount: float  # 历史累计总金额
-    amount: float  # 平均金额（保持向后兼容）
+    total_amount: Decimal  # 历史累计总金额
+    amount: Decimal  # 平均金额（保持向后兼容）
     frequency: int  # 周期天数，如30表示每30天一次
-    confidence_score: float  # 0-100的置信度分数
+    confidence_score: Decimal  # 0-100的置信度分数
     last_occurrence: str  # 最近一次发生日期
     count: int  # 识别到的交易次数
     combination_key: str  # 完整的组合键，用于精确匹配
@@ -64,14 +65,14 @@ class RecurringExpense:
 class ExpenseTrend:
     """支出趋势数据点"""
     date: str
-    value: float
+    value: Decimal
     category: str = "total"  # 可以是 'total', 'recurring', 'flexible'
 
 @dataclass
 class ExpenseAnalysisData:
     """支出分析综合数据"""
     target_month: str
-    total_expense: float
+    total_expense: Decimal
     expense_trend: List[ExpenseTrend]  # 近6个月趋势
     recurring_expenses: List[RecurringExpense]  # 周期性支出排行
     flexible_composition: List[CompositionItem]  # 弹性支出分类占比
